+++ banyand/tsdb/seriesdb.go
@@ -512,7 +512,7 @@ func newSeriesDataBase(ctx context.Context, shardID common.ShardID, path string,
 	if o != nil {
 		options := o.(DatabaseOpts)
 		if options.SeriesMemSize > 1 {
-			memSize = options.SeriesMemSize
+			memSize = int64(options.SeriesMemSize)
 		} else {
 			memSize = defaultKVMemorySize
 		}
